use auth middleware for create, update and delete posts

+++ routes/post.js
@@ -1,17 +1,17 @@
 const express = require("express");
 const router = express.Router();
 
-const yup = require('yup')
+const yup = require('yup');
+const auth = require("../middlewares/auth");
 
 const Post = require('./../models/Posts')
 
 
 
-router.post('/', async (req, res) => {
+router.post('/', auth, async (req, res) => {
   try {
     await yup.object({
       title: yup.string().required(),
-      user: yup.string().required(),
       body: yup.string().required()
     })
       .validate(req.body)
@@ -23,7 +23,11 @@ router.post('/', async (req, res) => {
     return res.status(406).json({ error: 'Title already exists' })
   }
 
-  const post = await Post.create({ ...req.body, slug: req.body.title.toLowerCase().replaceAll(' ', '-') })
+  const post = await Post.create({
+    ...req.body,
+    slug: req.body.title.toLowerCase().replaceAll(' ', '-'),
+    user: req.user.id
+  })
 
   return res.status(201).json(post)
 
@@ -37,7 +41,7 @@ router.get('/:slug', async (req, res) => {
   return res.json(await Post.findOne({ slug: req.params.slug }).populate('user', 'name').exec())
 })
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', auth, async (req, res) => {
   try {
     await yup.object({
       title: yup.string(),
@@ -49,14 +53,31 @@ router.put('/:id', async (req, res) => {
     return res.json(e.errors)
   }
 
+  const post = await Post.findById(req.params.id).exec()
+
+  if (post.user !== req.user.id) {
+    return res.status(403).json({
+      message: "You are not allowed to update this post"
+    })
+  }
+
   await Post.findByIdAndUpdate(req.params.id, req.body).exec()
   return res.status(200).json({
     message: "Post updated successfully"
   })
 })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
+  const post = await Post.findById(req.params.id).exec()
+
+  if (post.user !== req.user.id) {
+    return res.status(403).json({
+      message: "You are not allowed to delete this post"
+    })
+  }
+
   await Post.findByIdAndDelete(req.params.id).exec()
+
   return res.status(204).send()
 })
